internal: move device, telemetron packages to spacex_api
Regenerating protobuf sources with newer Starlink dishy software
results in a spacex_api package being used for the device and
telemetron API packages.

This moves the existing files before regenerating, in order to
keep diffs between the file revisions.

I am not sure why Starlink seemingly made this change, it is
inconvenient, so I may figure out a way to merge it back into the
spacex/api package at a later date.
